[BT] SP-490 #close #time 1h #comment menu 와 direct chat 간에 one to many 속성 중 mapping 으로 인해 remove 버그가 발견되어 fix함.  [RV] +review SR @admin

+++ standard/project/web/src/main/java/egovframework/api/rivalWar/menu/controller/AdminMenuController.java
@@ -106,7 +106,16 @@ public class AdminMenuController extends GenericAbstractController {
         if (bindingResult.hasErrors())
             throw new RuntimeException();
 
-        jsTreeHibernateDTO.setStatus(menuService.removeNode(jsTreeHibernateDTO));
+        MenuDTO targetMenuNode = menuService.getNode(jsTreeHibernateDTO);
+
+        HashSet<DirectChatDTO> dummyChatDTOs = new HashSet<DirectChatDTO>();
+        targetMenuNode.setDirectChatDTOs(dummyChatDTOs);
+        if ( menuService.alterNode(targetMenuNode) == 1 ){
+            jsTreeHibernateDTO.setStatus(menuService.removeNode(jsTreeHibernateDTO));
+        }else{
+            throw new RuntimeException("none remove target node");
+        }
+
         setJsonDefaultSetting(jsTreeHibernateDTO);
 
         ModelAndView modelAndView = new ModelAndView("jsonView");
